Cyclopedia: signed defense breakdown + skip zero component rows
- Defence "From Skill" can now be negative: read defenseEquipment /
  shieldingSkill / defenseWheel as signed int16 (struct + parser), so the
  skill's contribution (final - equipment) renders as e.g. -8.
- renderAndSkipZeroStats now skips zero-value component rows too, and skips a
  section's component rows when its header is skipped. Fixes the Dodge / Magic
  Shield Capacity sub-rows (From Amplification, Flat, Percent, extra From
  Base/Wheel) bleeding into the Mitigation list when those headers were 0.

+++ src/client/protocolgameparse.cpp
@@ -6133,10 +6133,10 @@ void ProtocolGame::parseCyclopediaCharacterInfo(const InputMessagePtr& msg)
             }
 
             data.defense = msg->getU16();
-            data.defenseEquipment = msg->getU16();
+            data.defenseEquipment = static_cast<int16_t>(msg->getU16()); // signed: may be negative
             data.defenseSkillType = msg->getU8();
-            data.shieldingSkill = msg->getU16();
-            data.defenseWheel = msg->getU16();
+            data.shieldingSkill = static_cast<int16_t>(msg->getU16()); // signed: final - equipment, may be negative
+            data.defenseWheel = static_cast<int16_t>(msg->getU16()); // signed
             msg->getU16(); // unused
 
             data.mitigation = msg->getDouble();
